This is the inital Commit
This is the first commit and adds some of the groudwork for the
networking behind the project. This includes getting urls and
previewing them. Currently the previewing is very primative
and could be improved so that only the video shows. The code
is also messing and could do with a clean up.